move auth logic to AuthToken module
