+++ back/app/controllers/omniauth_callback_controller.rb
@@ -154,7 +154,7 @@ class OmniauthCallbackController < ApplicationController
       { sub: entity.id }
     end
 
-    AuthToken.new payload: payload.merge({
+    AuthToken::AuthToken.new payload: payload.merge({
       provider: provider,
       logout_supported: AuthenticationService.new.supports_logout?(provider)
     })
