+++ back/engines/commercial/id_bosa_fas/spec/requests/bosa_fas_verification_spec.rb
@@ -6,7 +6,7 @@ require 'rspec_api_documentation/dsl'
 describe 'bosa_fas verification' do
   before do
     @user = create(:user, first_name: 'Rudolphi', last_name: 'Raindeari')
-    @token = AuthToken.new(payload: @user.to_token_payload).token
+    @token = AuthToken::AuthToken.new(payload: @user.to_token_payload).token
     OmniAuth.config.test_mode = true
     OmniAuth.config.mock_auth[:bosa_fas] = OmniAuth::AuthHash.new({
       'provider' => 'bosa_fas',
