+++ back/engines/commercial/public_api/app/controllers/public_api/public_api_controller.rb
@@ -2,7 +2,7 @@
 
 module PublicApi
   class PublicApiController < ActionController::API
-    include ::Authenticable
+    include ::AuthToken::Authenticable
     include Pundit
 
     before_action :authenticate_api_client
