+++ back/engines/commercial/public_api/spec/acceptance/phases_spec.rb
@@ -6,7 +6,7 @@ require 'rspec_api_documentation/dsl'
 resource 'Phases' do
   before do
     api_token = PublicApi::ApiClient.create
-    token = AuthToken.new(payload: api_token.to_token_payload).token
+    token = AuthToken::AuthToken.new(payload: api_token.to_token_payload).token
     header 'Authorization', "Bearer #{token}"
   end
