+++ back/engines/commercial/public_api/spec/acceptance/projects_spec.rb
@@ -6,7 +6,7 @@ require 'rspec_api_documentation/dsl'
 resource 'Projects' do
   before do
     api_token = PublicApi::ApiClient.create
-    token = AuthToken.new(payload: api_token.to_token_payload).token
+    token = AuthToken::AuthToken.new(payload: api_token.to_token_payload).token
     header 'Authorization', "Bearer #{token}"
   end
